Initial commit for S4
A simple tool that will load as much as possible to an S3 target bucket. Note that there are at current, no goroutines used explicitly. This is because the design of Go is such that the runtime *should* create a new "M" after it finds a blocking "G", and thus hides all of this for us in the background.